Cambios  de rutas y mejorar la semantica
