+++ src/components/atoms/GenerarCodigo.jsx
