+++ src/components/organisms/Nav.astro
@@ -1,7 +1,7 @@
 ---
-import IconsHerf from "./IconsHref.astro";
-import Saludar from "./Saludar.jsx";
-import { ShieldAlt } from "../icons/AllIcons";
+import IconsHerf from "@/components/atoms/IconsHref.astro";
+import Saludar from "@/components/atoms/Saludar.jsx";
+import { ShieldAlt } from "@/icons/AllIcons.tsx";
 ---
 
 <nav class="flex-1 max-w-7xl mx-auto p-4">
