+++ src/layouts/Layout.astro
@@ -1,7 +1,7 @@
 ---
-import Nav from "../components/Nav.astro";
-import Footer from "../components/Footer.astro";
-import "../styles/global.css";
+import Nav from "@/components/organisms/Nav.astro";
+import Footer from "@/components/atoms/Footer.astro";
+import "@/styles/global.css";
 ---
 
 <!doctype html>
