+++ src/pages/PreguntasFrecuentes.astro
@@ -1,5 +1,5 @@
 ---
-import Layout from "../layouts/Layout.astro";
+import Layout from "@/layouts/Layout.astro";
 ---
 
 <Layout>
@@ -9,7 +9,7 @@ import Layout from "../layouts/Layout.astro";
     </h1>
 
     <div class="space-y-6">
-      <div>
+      <article>
         <h2 class="text-xl font-semibold text-black">
           ¿Puedo hacer una denuncia de forma completamente anónima?
         </h2>
@@ -18,9 +18,9 @@ import Layout from "../layouts/Layout.astro";
           tu identidad. No pedimos nombres, correos, ni datos personales. Puedes
           denunciar sin temor.
         </p>
-      </div>
+      </article>
 
-      <div>
+      <article>
         <h2 class="text-xl font-semibold text-black">
           ¿Cómo se protege mi informaci&oacute;?
         </h2>
@@ -29,9 +29,9 @@ import Layout from "../layouts/Layout.astro";
           pueda ser le&iacute;da por personal autorizado. Los datos viajan
           cifrados desde tu navegador hasta nuestra base de datos.
         </p>
-      </div>
+      </article>
 
-      <div>
+      <article>
         <h2 class="text-xl font-semibold text-black">
           ¿Qué tipo de denuncias puedo enviar?
         </h2>
@@ -40,9 +40,9 @@ import Layout from "../layouts/Layout.astro";
           violencia, discriminaci&oacute;n, acoso, negligencia, entre otros.
           Cualquier situación grave puede ser reportada.
         </p>
-      </div>
+      </article>
 
-      <div>
+      <article>
         <h2 class="text-xl font-semibold text-black">
           ¿Qui&eacute;n recibe mi denuncia?
         </h2>
@@ -51,9 +51,9 @@ import Layout from "../layouts/Layout.astro";
           únicamente mediante una clave privada que descifra la
           informaci&oacute;n.
         </p>
-      </div>
+      </article>
 
-      <div>
+      <article>
         <h2 class="text-xl font-semibold text-black">
           ¿Puedo recibir una respuesta o seguimiento?
         </h2>
@@ -62,7 +62,7 @@ import Layout from "../layouts/Layout.astro";
           decides no dejar nada, tu denuncia seguir&aacute; siendo recibida y
           tratada con seriedad.
         </p>
-      </div>
+      </article>
     </div>
   </section>
 </Layout>
