+++ src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
-import Layout from "../layouts/Layout.astro";
-import "../styles/global.css";
+import Layout from "@/layouts/Layout.astro";
+import "@/styles/global.css";
 import {
   UserShield,
   Lock,
@@ -8,20 +8,20 @@ import {
   UserSecret,
   Circule,
   Sign,
-} from "../icons/AllIcons";
+} from "@/icons/AllIcons.tsx";
 
-import GeneradorCodigo from "../components/GenerarCodigo.jsx";
-import AccederConCodigo from "../components/AccederCodigo";
+import GeneradorCodigo from "@/components/atoms/GenerarCodigo.jsx";
+import AccederConCodigo from "@/components/moleculas/AccederCodigo";
 
 // Welcome to Astro! Wondering what to do next? Check out the Astro documentation at https://docs.astro.build
 // Don't want to use any of this? Delete everything in this file, the `assets`, `components`, and `layouts` directories, and start fresh.
 ---
 
 <Layout>
-  <main
+  <div
     class="container mx-auto flex flex-col items-center h-full justify-around"
   >
-    <div
+    <section
       class="lg:max-w-6xl md:max-w-3xl bg-white p-8 rounded-lg shadow-lg shadow-zinc-300"
     >
       <h1
@@ -67,9 +67,9 @@ import AccederConCodigo from "../components/AccederCodigo";
           <AccederConCodigo client:load />
         </div>
       </div>
-    </div>
+    </section>
 
-    <div class="mt-10 sm:mt-8 md:mt-7 lg:mt-8 lg:max-w-6xl md:max-w-3xl">
+    <section class="mt-10 sm:mt-8 md:mt-7 lg:mt-8 lg:max-w-6xl md:max-w-3xl">
       <h2 class="text-3xl font-bold mb-6 text-center">
         ¿Por qué usar este sistema?
       </h2>
@@ -83,7 +83,7 @@ import AccederConCodigo from "../components/AccederCodigo";
         class="mt-12 text-center flex flex-col md:flex-row items-stretch justify-center gap-6"
       >
         <!-- Card 1 -->
-        <div
+        <article
           class="w-full md:w-1/3 p-4 flex flex-col items-center bg-white rounded-lg shadow-lg shadow-gray-300 hover:-translate-y-2 hover:scale-100 duration-400"
         >
           <div
@@ -98,17 +98,17 @@ import AccederConCodigo from "../components/AccederCodigo";
             Este sistema es una iniciativa de la comunidad para promover la
             transparencia y la justicia.
           </p>
-        </div>
+        </article>
 
         <!-- Card 2 -->
         <div
           class="w-full md:w-1/3 p-4 flex flex-col items-center bg-white rounded-lg shadow-lg shadow-gray-300 hover:-translate-y-2 hover:scale-100 duration-400"
         >
-          <div
+          <article
             class="w-20 h-20 bg-[#1e56a0] rounded-full flex items-center justify-center"
           >
             <Lock className="w-12 h-12 text-white" />
-          </div>
+          </article>
           <h3 class="mt-3 text-lg font-bold text-[#1e56a0]">Seguridad</h3>
           <p class="text-gray-500 mt-2">
             Todas las denuncias son tratadas con la máxima confidencialidad y
@@ -117,7 +117,7 @@ import AccederConCodigo from "../components/AccederCodigo";
         </div>
 
         <!-- Card 3 -->
-        <div
+        <article
           class="w-full md:w-1/3 p-4 flex flex-col items-center bg-white rounded-lg shadow-lg shadow-gray-300 hover:-translate-y-2 hover:scale-100 duration-400"
         >
           <div
@@ -130,8 +130,8 @@ import AccederConCodigo from "../components/AccederCodigo";
             Nuestro objetivo es fomentar un entorno de confianza y transparencia
             en la comunidad.
           </p>
-        </div>
+        </article>
       </div>
-    </div>
-  </main>
+    </section>
+  </div>
 </Layout>
