+++ tsconfig.json
@@ -8,6 +8,12 @@
     "dist"
   ],
   "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@/*": [
+        "src/*"
+      ]
+    },
     "jsx": "react-jsx",
     "jsxImportSource": "react"
   }
